+++ src/Mdbx.hs
@@ -6,20 +6,21 @@ Maintainer  : fjvallarino@gmail.com
 Stability   : experimental
 Portability : non-portable
 
-Main module for libmdbx-hs. This is what most applications should import.
+Bindings for libmdbx, a high-performance, in-process, key-value store.
 
-In case you only need to store, update and retrieve data, check the
-"Mdbx.Database" module.
+This is the module most applications should import.
 
-You will also want to check "Mdbx.Types" in order to be able to store your data
-types. In general, you will also want to use a serialization library such as
-<https://hackage.haskell.org/package/store store> or
-<https://hackage.haskell.org/package/cereal cereal>.
+For the high level API, check the "Mdbx.Database" module.
 
-If you want fine grained control or using cursors, check "Mdbx.API".
+For the low level API, or if you need finer grained control/using cursors, check
+"Mdbx.API".
 
-The "Mdbx.FFI" (not exported by this module) provides direct, low level,
-bindings to libmdbx.
+You will also want to check "Mdbx.Types" to be able to store your data types. In
+general, you will also want to use a serialization library such as
+<https://hackage.haskell.org/package/store store>.
+
+"Mdbx.FFI" (not exported by this module) provides direct, low level, bindings to
+libmdbx.
 -}
 module Mdbx (
   module Mdbx.API,